# docs/swagger.yaml
basePath: /
definitions:
  models.AnimeDetailResponse:
    properties:
      anime_slug:
        type: string
      confidence_score:
        type: number
      cover:
        type: string
      details:
        $ref: '#/definitions/models.AnimeDetails'
      episode_list:
        items:
          $ref: '#/definitions/models.EpisodeListItem'
        type: array
      genre:
        items:
          type: string
        type: array
      judul:
        type: string
      message:
        type: string
      penonton:
        type: string
      rating:
        $ref: '#/definitions/models.AnimeRating'
      recommendations:
        items:
          $ref: '#/definitions/models.RecommendationItem'
        type: array
      sinopsis:
        type: string
      skor:
        type: string
      source:
        type: string
      status:
        type: string
      tipe:
        type: string
      url:
        type: string
    type: object
  models.AnimeDetails:
    properties:
      Duration:
        type: string
      English:
        type: string
      Japanese:
        type: string
      Producers:
        type: string
      'Released:':
        type: string
      Season:
        type: string
      Source:
        type: string
      Status:
        type: string
      Studio:
        type: string
      Total Episode:
        type: string
      Type:
        type: string
    type: object
...
  models.AnimeTerbaruResponse:
    properties:
      confidence_score:
        type: number
      data:
        items:
          $ref: '#/definitions/models.AnimeTerbaruItem'
        type: array
      message:
        type: string
      source:
        type: string
    type: object
  models.DayScheduleResponse:
    properties:
      confidence_score:
        type: number
      data:
        items:
          $ref: '#/definitions/models.ScheduleItem'
        type: array
      message:
        type: string
      source:
        type: string
    type: object
...
  models.EpisodeDetailResponse:
    properties:
      anime_info:
        $ref: '#/definitions/models.AnimeInfo'
      confidence_score:
        type: number
      download_links:
        $ref: '#/definitions/models.DownloadLinksGroup'
      message:
        type: string
      navigation:
        $ref: '#/definitions/models.EpisodeNavigation'
      other_episodes:
        items:
          $ref: '#/definitions/models.OtherEpisode'
        type: array
      release_info:
        type: string
      source:
        type: string
      streaming_servers:
        items:
          $ref: '#/definitions/models.StreamingServer'
        type: array
      thumbnail_url:
        type: string
      title:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      confidence_score:
        type: number
      error:
        type: boolean
      message:
        type: string
    type: object
  models.HomeResponse:
    properties:
      confidence_score:
        type: number
      jadwal_rilis:
        $ref: '#/definitions/models.ScheduleData'
      message:
        type: string
      movies:
        items:
          $ref: '#/definitions/models.MovieItem'
        type: array
      new_eps:
        items:
          $ref: '#/definitions/models.NewEpisodeItem'
        type: array
      source:
        type: string
      top10:
        items:
          $ref: '#/definitions/models.Top10Item'
        type: array
    type: object
...
  models.MovieResponse:
    properties:
      confidence_score:
        type: number
      data:
        items:
          $ref: '#/definitions/models.MovieDetailItem'
        type: array
      message:
        type: string
      source:
        type: string
    type: object
...
  models.ScheduleResponse:
    properties:
      confidence_score:
        type: number
      data:
        $ref: '#/definitions/models.ScheduleData'
      message:
        type: string
      source:
        type: string
    type: object
  models.SearchResponse:
    properties:
      confidence_score:
        type: number
      data:
        items:
          $ref: '#/definitions/models.SearchResultItem'
        type: array
      message:
        type: string
      source:
        type: string
    type: object
...
  repository.AnimeDetailResponse:
    properties:
      confidence_score:
        example: 1
        type: number
      data:
        $ref: '#/definitions/repository.AnimeDetailData'
      message:
        type: string
      source:
        type: string
    type: object
...
  repository.AnimeTerbaruResponse:
    properties:
      confidence_score:
        example: 1
        type: number
      data:
        items:
          $ref: '#/definitions/repository.AnimeTerbaruItem'
        type: array
      message:
        type: string
      source:
        type: string
    type: object
...
  repository.EpisodeDetailResponse:
    properties:
      confidence_score:
        example: 1
        type: number
      data:
        $ref: '#/definitions/repository.EpisodeDetailData'
      message:
        type: string
      source:
        type: string
    type: object
...
  repository.FinalResponse:
    properties:
      confidence_score:
        type: number
      data:
        $ref: '#/definitions/repository.HomeData'
      message:
        type: string
      source:
        type: string
    type: object
...
  repository.MovieListResponse:
    properties:
      confidence_score:
        example: 1
        type: number
      data:
        items:
          $ref: '#/definitions/repository.MovieItem'
        type: array
      message:
        type: string
      source:
        type: string
    type: object
...
  repository.SearchResponse:
    properties:
      confidence_score:
        example: 1
        type: number
      data:
        items:
          $ref: '#/definitions/repository.SearchResultItem'
        type: array
      message:
        type: string
      source:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A robust API fallback service for anime content with automatic failover
    capabilities
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: API Fallback Service
  version: "1.0"
paths:
  /api/v1/anime-detail:
    get:
      consumes:
      - application/json
      description: Mengambil detail anime, film, atau series termasuk episode, sinopsis,
        dan rekomendasi. Slug dapat berupa 'nama-anime', 'film/nama-film', atau 'series/nama-series'
      parameters:
      - description: 'Anime/Movie/Series slug (contoh: ''kobane-2022'', ''film/kobane-2022'',
          ''series/legend-of-the-female-general'')'
        in: query
        name: anime_slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AnimeDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get anime/movie/series detail
      tags:
      - Detail
  /api/v1/anime-detail/:
    get:
      consumes:
      - application/json
      description: Mengambil detail lengkap sebuah anime berdasarkan slug seri atau
        slug episode.
      parameters:
      - description: Slug dari anime yang ingin dicari
        in: query
        name: anime_slug
        required: true
        type: string
      - description: Force refresh cache (opsional, belum diimplementasikan)
        in: query
        name: force_refresh
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Detail anime berhasil diambil
          schema:
            $ref: '#/definitions/repository.AnimeDetailResponse'
        "400":
          description: Parameter anime_slug tidak ditemukan
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Anime tidak ditemukan
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Anime Detail
      tags:
      - Anime Detail
  /api/v1/anime-terbaru:
    get:
      consumes:
      - application/json
      description: Mengambil daftar anime terbaru dengan pagination
      parameters:
      - default: 1
        description: Nomor halaman
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AnimeTerbaruResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get anime terbaru
      tags:
      - Anime
  /api/v1/anime-terbaru/:
    get:
      description: Mengambil daftar rilis anime terbaru dengan paginasi.
      parameters:
      - default: 1
        description: Nomor halaman
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Daftar rilis terbaru berhasil diambil
          schema:
            $ref: '#/definitions/repository.AnimeTerbaruResponse'
        "400":
          description: Parameter halaman tidak valid
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Latest Anime Releases
      tags:
      - Anime List
  /api/v1/episode-detail:
    get:
      consumes:
      - application/json
      description: Mengambil detail episode termasuk server streaming dan link download
      parameters:
      - description: URL episode
        in: query
        name: episode_url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EpisodeDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get episode detail
      tags:
      - Detail
  /api/v1/episode-detail/:
    get:
      consumes:
      - application/json
      description: Mengambil detail lengkap sebuah episode berdasarkan URL.
      parameters:
      - description: URL lengkap dari halaman episode
        in: query
        name: episode_url
        required: true
        type: string
      - description: Force refresh cache (opsional, belum diimplementasikan)
        in: query
        name: force_refresh
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Detail episode berhasil diambil
          schema:
            $ref: '#/definitions/repository.EpisodeDetailResponse'
        "400":
          description: Parameter episode_url tidak valid atau kosong
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Episode tidak ditemukan
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Episode Detail
      tags:
      - Episode Detail
  /api/v1/home:
    get:
      consumes:
      - application/json
      description: Mengambil data homepage termasuk top 10 anime, episode terbaru,
        film terbaru, dan jadwal rilis
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HomeResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get homepage data
      tags:
      - Homepage
  /api/v1/home/:
    get:
      consumes:
      - application/json
      description: Mengambil daftar anime terbaru, film, top 10, dan jadwal rilis
        mingguan.
      produces:
      - application/json
      responses:
        "200":
          description: Data berhasil diambil
          schema:
            $ref: '#/definitions/repository.FinalResponse'
        "500":
          description: Error internal server
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Latest Anime & Schedule
      tags:
      - Anime
  /api/v1/jadwal-rilis:
    get:
      consumes:
      - application/json
      description: Mengambil jadwal rilis anime per hari
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ScheduleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get jadwal rilis
      tags:
      - Schedule
  /api/v1/jadwal-rilis/:
    get:
      consumes:
      - application/json
      description: Mengambil jadwal rilis anime untuk semua hari.
      parameters:
      - description: Force refresh cache (opsional, belum diimplementasikan)
        in: query
        name: force_refresh
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Jadwal rilis berhasil diambil
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error internal server
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Release Schedule
      tags:
      - Jadwal Rilis
  /api/v1/jadwal-rilis/{day}:
    get:
      consumes:
      - application/json
      description: Mengambil jadwal rilis anime untuk hari tertentu
      parameters:
      - description: Nama hari (monday, tuesday, wednesday, thursday, friday, saturday,
          sunday)
        in: path
        name: day
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DayScheduleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get jadwal rilis by day
      tags:
      - Schedule
  /api/v1/movie:
    get:
      consumes:
      - application/json
      description: Mengambil daftar film dengan pagination
      parameters:
      - default: 1
        description: Nomor halaman
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MovieResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get movies
      tags:
      - Movies
  /api/v1/movie/:
    get:
      consumes:
      - application/json
      description: Mengambil daftar anime dari halaman utama (latest) sebagai pengganti
        daftar film.
      parameters:
      - default: 1
        description: Nomor halaman
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Daftar berhasil diambil
          schema:
            $ref: '#/definitions/repository.MovieListResponse'
        "400":
          description: Parameter halaman tidak valid
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error internal server
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get "Movie" List (from Latest)
      tags:
      - Movie
  /api/v1/search:
    get:
      consumes:
      - application/json
      description: Mencari anime berdasarkan judul
      parameters:
      - description: Query pencarian
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Search anime
      tags:
      - Search
  /api/v1/search/:
    get:
      description: Mencari anime berdasarkan query.
      parameters:
      - description: Kata kunci pencarian
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Hasil pencarian
          schema:
            $ref: '#/definitions/repository.SearchResponse'
        "400":
          description: Parameter query tidak ditemukan
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search Anime
      tags:
      - Anime List
...
schemes:
- http
- https
swagger: "2.0"
...